[testsuite] Breakdown Aptos CLI tests into different files
